+++ src/app/app.routes.ts
@@ -6,6 +6,10 @@ import { LoginComponent } from './login/login.component';
 import { BetComponent } from './bet/bet.component';
 import { LiveComponent } from './live/live.component';
 import { loggedInGuard } from './logged-in.guard';
+import { PendingRacesComponent } from './races/pending-races/pending-races.component';
+import { FinishedRacesComponent } from './races/finished-races/finished-races.component';
+import { racesResolver } from './races.resolver';
+import { raceResolver } from './race.resolver';
 
 export const routes: Routes = [
   { path: '', component: HomeComponent },
@@ -15,13 +19,40 @@ export const routes: Routes = [
     path: 'races',
     canActivate: [loggedInGuard],
     children: [
-      { path: '', component: RacesComponent },
+      {
+        path: '',
+        component: RacesComponent,
+        children: [
+          { path: '', pathMatch: 'full', redirectTo: 'pending' },
+          {
+            path: 'pending',
+            component: PendingRacesComponent,
+            resolve: {
+              races: racesResolver
+            }
+          },
+          {
+            path: 'finished',
+            component: FinishedRacesComponent,
+            resolve: {
+              races: racesResolver
+            }
+          }
+        ]
+      },
       {
         path: ':raceId',
-        children: [
-          { path: '', component: BetComponent },
-          { path: 'live', component: LiveComponent }
-        ]
+        component: BetComponent,
+        resolve: {
+          race: raceResolver
+        }
+      },
+      {
+        path: ':raceId/live',
+        component: LiveComponent,
+        resolve: {
+          race: raceResolver
+        }
       }
     ]
   }