+++ src/app/race.service.ts
@@ -3,7 +3,7 @@ import { LiveRaceModel, RaceModel } from './models/race.model';
 import { Observable, map, takeWhile } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '../environments/environment';
-import { FINISHED_RACE_STATUS, PENDING_RACE_STATUS } from '../constants/status';
+import { FINISHED_RACE_STATUS } from '../constants/status';
 import { PonyWithPositionModel } from './models/pony.model';
 import { WsService } from './ws.service';
 
@@ -16,8 +16,8 @@ export class RaceService {
     private wsService: WsService
   ) {}
 
-  list(): Observable<Array<RaceModel>> {
-    return this.http.get<Array<RaceModel>>(`${environment.baseUrl}/api/races`, { params: { status: PENDING_RACE_STATUS } });
+  list(status: 'PENDING' | 'RUNNING' | 'FINISHED'): Observable<Array<RaceModel>> {
+    return this.http.get<Array<RaceModel>>(`${environment.baseUrl}/api/races`, { params: { status } });
   }
 
   get(raceId: number): Observable<RaceModel> {